% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonCorrelationModel.R
\name{compareICC}
\alias{compareICC}
\title{Compare intraclass correlation between two groups}
\usage{
compareICC(chisq_res, method = "auto", print.out = TRUE)
}
\arguments{
\item{chisq_res}{a dataframe, the result of the twin.chisqTest function.}

\item{method}{character, methods for parameter estimation (e.g. "auto","normal","gof"). Default is "auto".
Recommended method is 'normal' when total number of monochorionic sample >=100 or dichorionic sample >=100.
#' Recommended method is 'gof' when total number of monochorionic sample <100 or dichorionic sample <100.}

\item{print.out}{a boolean, Wwether or not to print results.}
}
\value{
Return a data frame. Rows are genes/drivers, columns are "ID", "logFC", "AveExpr", "t", "P.Value", "adj.P.Val", "B", "Z-statistics", "Ave.G1" and "Ave.G0".
Names of the columns may vary from different group names. Sorted by P-values.
}
\description{
\code{compareICC} is a function to report CIs for the ICCs using a goodness-of fit
approach and tested whether the 2 ICCs differ.
}
\examples{
data <- load_example_data();
chisq_res <- twin.chisqTest(data = data,famid = 'FAMID',outcome = 'BPD',zyg = 'zygo')
compareICC(chisq_res$stat_mat)

}
