% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratificationOR.R
\name{compareMD}
\alias{compareMD}
\title{Estimate of the familial association for disease}
\usage{
compareMD(result.twin.index, print.out = TRUE)
}
\arguments{
\item{result.twin.index}{a list, result of function twin.index.}

\item{print.out}{a boolean, Whether or not to print results.}
}
\value{
Return two dataframe. 
OR is different effects odds ratios.
test is different test statistics for the common odds ratio.
}
\description{
\code{compareMD} is a function for analyzing dichotomous outcomes in unselected 
samples of twin pairs using stratified estimators of the odds ratio.
}
\examples{
data <- load_example_data();
twin.index_res <- twin.index(data=data,famid='FAMID',outcome='BPD',zyg='zygo')
compareMD(twin.index_res)

}
