% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_function.R
\name{load_example_data}
\alias{load_example_data}
\title{Load example data}
\usage{
load_example_data()
}
\value{
The loaded example data
}
\description{
This function loads example data for testing and learning purposes.
}
\examples{
data <- load_example_data()
}
