% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{logisticMD}
\alias{logisticMD}
\title{Estimate of the familial association for disease}
\usage{
logisticMD(
  data,
  famid,
  outcome,
  outcome_present = 1,
  outcome_absent = 0,
  zyg,
  Monochorionic_value = 1,
  Dichorionic_value = 0,
  seed = 42,
  print.out = TRUE
)
}
\arguments{
\item{data}{a dataframe, raw data}

\item{famid}{character, family Number.}

\item{outcome}{character, outcome variable.}

\item{outcome_present}{character or numeric, present in outcome variable. Default is 1.}

\item{outcome_absent}{character or numeric, absent in outcome variable. Default is 0.}

\item{zyg}{character, indicates the type of twin, in the example data 1 is monochorionic and 0 is dichorionic.}

\item{Monochorionic_value}{character or numeric, the value indicating monochorionic twins. Default is 1.}

\item{Dichorionic_value}{character or numeric, the value indicating dichorionic twins. Default is 0.}

\item{seed}{numeric, the methodology will use this seed to randomly designating one member of each twin
pair as an “index” twin and the other member as the “co-twin.”}

\item{print.out}{a boolean, Whether or not to print results.}
}
\value{
Return a dataframe 
The result of logistic regression method.
}
\description{
\code{logisticMD} is a function for analyzing  dichotomous outcomes in unselected 
samples of twins using logistic regression based on unconditional
maximum likelihood estimation.
}
\examples{
data <- load_example_data();
logisticMD(data=data,famid='FAMID',outcome='BPD',zyg='zygo')

}
