% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metsMD.R
\name{metsMD}
\alias{metsMD}
\title{Analysis of bivariate binomial data}
\usage{
metsMD(
  data,
  famid,
  outcome,
  outcome_present = 1,
  outcome_absent = 0,
  zyg,
  Monochorionic_value = 1,
  Dichorionic_value = 0,
  cov_var = NULL,
  use_model = "bp",
  type = "ace",
  print.out = TRUE
)
}
\arguments{
\item{data}{a dataframe, raw data}

\item{famid}{character, family Number.}

\item{outcome}{character, outcome variable.}

\item{outcome_present}{character or numeric, present in outcome variable. Default is 1.}

\item{outcome_absent}{character or numeric, absent in outcome variable. Default is 0.}

\item{zyg}{character, indicates the type of twin, in the example data 1 is monochorionic and 0 is dichorionic.}

\item{Monochorionic_value}{character or numeric, the value indicating monochorionic twins. Default is 1.}

\item{Dichorionic_value}{character or numeric, the value indicating dichorionic twins. Default is 0.}

\item{cov_var}{vector, the covariant variables.}

\item{use_model}{character, fit the model: Pairwise odds ratio model (or), Bivariate Probit model (bp), additive gamma random effects model (gamma). Default is "bp".}

\item{type}{character, With random effects (un) and special functionality for polygenic random effects modelling such as ACE (ace), ADE (ade), AE (ae) and so forth. Default is "ace".}

\item{print.out}{a boolean, Whether or not to print results.}
}
\value{
Return a class of summary.mets.twostage
}
\description{
\code{metsMD} is a function for analyzing bivariate binomial data with correct for some covariates
based on R package mets (https://kkholst.github.io/mets/index.html).
}
\examples{
data <- load_example_data();
metsMD(data=data,famid='FAMID',outcome='BPD',zyg='zygo',cov_var='RDS',use_model='bp')
metsMD(data=data,famid='FAMID',outcome='BPD',zyg='zygo',cov_var='RDS',use_model='bp',type='ade')
metsMD(data=data,famid='FAMID',outcome='BPD',zyg='zygo',cov_var='RDS',use_model='gamma',type='ace')

}
