% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_function.R
\name{twin.chisqTest}
\alias{twin.chisqTest}
\title{Sample size for counting monochorionic and dichorionic twins}
\usage{
twin.chisqTest(
  data,
  famid,
  outcome,
  outcome_present = 1,
  outcome_absent = 0,
  zyg,
  Monochorionic_value = 1,
  Dichorionic_value = 0,
  print.out = TRUE,
  ...
)
}
\arguments{
\item{data}{a dataframe, raw data}

\item{famid}{character, family id.}

\item{outcome}{character, outcome variable.}

\item{outcome_present}{character or numeric, present in outcome variable. Default is 1.}

\item{outcome_absent}{character or numeric, absent in outcome variable. Default is 0.}

\item{zyg}{character, indicates the type of twin, in the example data 1 is monochorionic and 0 is dichorionic.}

\item{Monochorionic_value}{character or numeric, the value indicating monochorionic twins. Default is 1.}

\item{Dichorionic_value}{character or numeric, the value indicating dichorionic twins. Default is 0.}

\item{print.out}{a boolean, Whether or not to print results.}
}
\value{
Return a list. Chisq_value and P_value are the results of the chi-square test and 
stat_mat is the cross-tabulation table.
}
\description{
\code{twin.chisqTest} is a function to count the difference in the number of monochorionic and dichorionic twins
}
\examples{
data <- load_example_data();
twin.chisqTest(data = data,famid = 'FAMID',outcome = 'BPD',zyg = 'zygo')

}
