% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_function.R
\name{twin.index}
\alias{twin.index}
\title{Organization of Twin Data for the Genetic Analysis of Dichotomous Outcomes}
\usage{
twin.index(
  data,
  famid,
  outcome,
  outcome_present = 1,
  outcome_absent = 0,
  zyg,
  Monochorionic_value = 1,
  Dichorionic_value = 0,
  seed = 42
)
}
\arguments{
\item{data}{a dataframe, raw data}

\item{famid}{character, family Number.}

\item{outcome}{character, outcome variable.}

\item{outcome_present}{character or numeric, present in outcome variable. Default is 1.}

\item{outcome_absent}{character or numeric, absent in outcome variable. Default is 0.}

\item{zyg}{character, indicates the type of twin, in the example data 1 is monochorionic and 0 is dichorionic.}

\item{Monochorionic_value}{character or numeric, the value indicating monochorionic twins. Default is 1.}

\item{Dichorionic_value}{character or numeric, the value indicating dichorionic twins. Default is 0.}

\item{seed}{numeric, the methodology will use this seed to randomly designating one member of each twin
pair as an “index” twin and the other member as the “co-twin.”}
}
\value{
Return a list. stat_MZ_mat and stat_DZ_mat are cross-tabulation tables.
}
\description{
\code{twin.index} is a function to count the the number of index twins and Co-twins
}
\examples{
data <- load_example_data();
twin.index(data = data,famid = 'FAMID',outcome = 'BPD',zyg = 'zygo')

}
