% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{twinstat_report}
\alias{twinstat_report}
\title{Generate twinstat report}
\usage{
twinstat_report(
  data,
  famid,
  outcome,
  outcome_present = 1,
  outcome_absent = 0,
  zyg,
  Monochorionic_value = 1,
  Dichorionic_value = 0,
  seed = 42,
  cov_var = NULL,
  filename = "./report.html"
)
}
\arguments{
\item{data}{a dataframe, raw data}

\item{famid}{character, family Number.}

\item{outcome}{character, outcome variable.}

\item{outcome_present}{character or numeric, present in outcome variable. Default is 1.}

\item{outcome_absent}{character or numeric, absent in outcome variable. Default is 0.}

\item{zyg}{character, indicates the type of twin, in the example data 1 is monochorionic and 0 is dichorionic.}

\item{Monochorionic_value}{character or numeric, the value indicating monochorionic twins. Default is 1.}

\item{Dichorionic_value}{character or numeric, the value indicating dichorionic twins. Default is 0.}

\item{seed}{numeric, see function [logisticMD()] and [GEE()].}

\item{cov_var}{vector, see function [metsMD()].}

\item{filename}{character, Markdown or HTML output file.  An HTML file
is specified using the .htm, .html, .HTM or .HTML file extension.
When html is specified, a similarly named markdown file is also
generated.}
}
\description{
\code{twinstat_report} this report will generate a report containing tables and graphs summarising the results.
}
\examples{
\dontrun{
data <- load_example_data()
data <- data \%>\% mutate_at(c("Sex","RDS","INST"),function(x)as.factor(x))
cov_var <- c("Sex","GA","BW","RDS","INST")
twinstat_report(data=data,famid='FAMID',outcome='BPD',zyg='zygo',cov_var=cov_var)
}
}
