% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratificationOR.R
\name{compareMD}
\alias{compareMD}
\title{Estimate of the familial association for disease}
\usage{
compareMD(result.twin.index, print.out = TRUE)
}
\arguments{
\item{result.twin.index}{a list, result of function twin.index.}

\item{print.out}{a boolean, Whether or not to print results.}
}
\value{
Return a dataframe. 
OR_MZ, OR_DZ, OR_common OR_a and OR_c are zygosity-specific odds ratios, common odds ratio, genetic effect and common environmental effects.
chisq_MH and p_value_MH are the Mantel-Haenszel test statistics for the common odds ratio.
chisq_G and  p_value_G are the Woolf-Haldane test statistics for the difference in the zygosity-specific odds ratios.
chisq_c and p_value_c are the test statistics for common environmental effects.
}
\description{
\code{compareMD} is a function for analyzing dichotomous outcomes in unselected 
samples of twin pairs using stratified estimators of the odds ratio.
}
\examples{
data <- load_example_data();
twin.index_res <- twin.index(data=data,famid='FAMID',outcome='BPD',zyg='zygo')
compareMD(twin.index_res)

}
